refactor: 팝오버 애니메이션 렌더링 로직 분리 [TM-127]
