+++ src/widgets/calendar/lib/usePopoverAnimation.ts
@@ -1,0 +1,17 @@
+import { useEffect } from 'react';
+
+export const usePopoverAnimation = (
+  isPopoverOpen: boolean,
+  setIsAnimating: React.Dispatch<React.SetStateAction<boolean>>,
+) => {
+  useEffect(() => {
+    if (!isPopoverOpen) {
+      setIsAnimating(true);
+      const timeout = setTimeout(() => {
+        setIsAnimating(false);
+      }, 250); // 애니메이션 지속 시간
+
+      return () => clearTimeout(timeout);
+    }
+  }, [isPopoverOpen, setIsAnimating]);
+};
